% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jblibminer-package.R
\docType{package}
\name{jblibminer-package}
\alias{jblibminer}
\alias{jblibminer-package}
\title{jblibminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jennybc/jblibminer}
  \item Report bugs at \url{https://github.com/jennybc/jblibminer/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

}
\keyword{internal}
